Replace modal with dedicated reset screen
